add comments && fix bug
