+++ internal/gf-cluster/cluster/client.go
@@ -12,10 +12,12 @@ import (
 var (
 	clients map[int64]*KubernetesClient
 )
+
 //KubernetesClient 是k8s访问的client
 type KubernetesClient struct {
 	ClientSet *kubernetes.Clientset
 }
+
 //CreateKubernetesClusterClient 通过k8s 配置信息创建client
 func CreateKubernetesClusterClient(data []byte) (*KubernetesClient, error) {
 	config, err := clientcmd.NewClientConfigFromBytes(data)
@@ -34,6 +36,7 @@ func CreateKubernetesClusterClient(data []byte) (*KubernetesClient, error) {
 
 	return &KubernetesClient{ClientSet: clientSet}, nil
 }
+
 //GetKubeClient 获取指定集群的client
 func GetKubeClient(kubernetesClusterId int64) (*KubernetesClient, error) {
 	client := clients[kubernetesClusterId]
